% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{NaiveBayes}
\alias{NaiveBayes}
\title{NaiveBayes}
\description{
trainable categorical Naive Bayes model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NaiveBayes-fit}{\code{NaiveBayes$fit()}}
\item \href{#method-NaiveBayes-compute_and_plot_importance}{\code{NaiveBayes$compute_and_plot_importance()}}
\item \href{#method-NaiveBayes-predict}{\code{NaiveBayes$predict()}}
\item \href{#method-NaiveBayes-predict_proba}{\code{NaiveBayes$predict_proba()}}
\item \href{#method-NaiveBayes-Print}{\code{NaiveBayes$Print()}}
\item \href{#method-NaiveBayes-Summary}{\code{NaiveBayes$Summary()}}
\item \href{#method-NaiveBayes-clone}{\code{NaiveBayes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$fit(
  X,
  y,
  preproc = TRUE,
  nb_classe = 6,
  epsilon = NULL,
  g_na = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A dataframe}

\item{\code{y}}{A vector}

\item{\code{preproc}}{A Boolean}

\item{\code{nb_classe}}{An integer}

\item{\code{epsilon}}{A float}

\item{\code{g_na}}{A Boolean

compute_and_plot_importance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-compute_and_plot_importance"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-compute_and_plot_importance}{}}}
\subsection{Method \code{compute_and_plot_importance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$compute_and_plot_importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A graph showing the importance of variables

predict
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-predict"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$predict(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{A dataframe}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
prediction on new_data

predict_proba
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$predict_proba(new_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{A dataframe}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of prediction probabilities
Print
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-Print"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-Print}{}}}
\subsection{Method \code{Print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$Print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string of model information

Summary
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-Summary"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-Summary}{}}}
\subsection{Method \code{Summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$Summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string with model details
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NaiveBayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
